refactor: divide functions onto different modules
